# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "WaveOne API",
        "contact": {
            "name": "Aneep Tandel",
            "url": "https://github.com/aneepct"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/advisory.{format?}": {
            "get": {
                "tags": [
                    "App Settings"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns app settings.",
                "operationId": "api.country.index",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App Settings."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/app-settings.{format?}": {
            "get": {
                "tags": [
                    "App Settings"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns app settings.",
                "operationId": "api.country.index",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App Settings."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/get-country.{format?}": {
            "get": {
                "tags": [
                    "Get Country"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns country overview.",
                "operationId": "api.country.index",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
    },
...
}